pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fibermorph"
version = "0.3.6"
description = "Toolkit for analyzing hair fiber morphology"
authors = ["tinalasisi <tina.lasisi@gmail.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/tinalasisi/fibermorph_update"
repository = "https://github.com/tinalasisi/fibermorph_update"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.4"
joblib = "^1.3.2"
pandas = "^2.2.0"
pillow = "^10.2.0"
requests = "^2.31.0"
matplotlib = "^3.8.2"
rawpy = "^0.19.0"
scipy = "^1.8"
shapely = "^2.0.2"
argparse = "^1.4.0"
tqdm = "^4.66.1"
scikit-image = "^0.22.0"
scikit-learn = "^1.4.0"
sympy = "^1.12"
pytest = "^8.0.0"
pyarrow = "^15.0.0"

[tool.poetry.scripts]
fibermorph = "fibermorph.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.0"
pre-commit = "^3.6.0"
